refactor(tests): introduce `MockMainMenu` and `MockEditor` for test simplification
- Replaced inline menu and editor mock implementations with `MockMainMenu` and `MockEditor` classes.
- Simplified `TestCore` by reusing modular reusable mock classes.
- Improved code maintainability by reducing redundancy in test setup.
